[platform] separate platform-related codes
