+++ src/platform/types.ts
@@ -1,0 +1,8 @@
+export interface WmpfProcessInfo {
+    pid: number;
+    version: number;
+}
+
+export interface IPlatform {
+    findWmpfProcess(): Promise<WmpfProcessInfo>;
+}
